requirements.txt
PyQt5==5.15.10
PyQt5-Qt5==5.15.11
PyQt5-sip==12.13.0
cryptography==41.0.7
python-dotenv==1.0.0
email-validator==2.1.0
requests==2.31.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

